HARMONY-1900: Refactored code.
